新建出库记录查询api以及完善出库页面

+++ server/src/main/java/com/abc/chenzeshenga/logistics/mapper/product/ProductOutWarehouseRecordMapper.java
@@ -18,5 +18,5 @@ public interface ProductOutWarehouseRecordMapper {
       @Param("productOutWarehouseRecord") ProductOutWarehouse productOutWarehouse,
       @Param("sqlLimit") SqlLimit sqlLimit);
 
-  int count(ProductOutWarehouse productOutWarehouse);
+  long count(ProductOutWarehouse productOutWarehouse);
 }
